# api/openapi.yaml
openapi: 3.1.0
info:
  title: memini
  description: A Kubernetes-ready memory service for AI agents.
  version: 0.1.0
  license:
    name: AGPL-3.0-only
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/memories:
    post:
      operationId: rememberMemory
      tags: [memory]
      summary: Remember (store) a memory
      parameters: [{ $ref: "#/components/parameters/Namespace" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RememberRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    get:
      operationId: listMemories
      tags: [memory]
      summary: List memories in a namespace (backs the admin UI browser)
      parameters:
        - { $ref: "#/components/parameters/Namespace" }
        - name: tier
          in: query
          description: Repeatable and/or comma-separated tier filter; omitted means all tiers.
          schema: { type: array, items: { $ref: "#/components/schemas/Tier" } }
        - name: level
          in: query
          description: Repeatable and/or comma-separated level filter; omitted means all levels.
          schema: { type: array, items: { $ref: "#/components/schemas/Level" } }
        - name: tag
          in: query
          description: Repeatable and/or comma-separated tag filter; a memory must carry every listed tag (AND).
          schema: { type: array, items: { type: string } }
        - name: meta
          in: query
          description: >
            Repeatable metadata filter in "key=value" form; a memory's top-level
            metadata must contain every listed pair (AND). Only string values match.
          schema: { type: array, items: { type: string } }
        - name: include_expired
          in: query
          schema: { type: boolean, default: false }
        - name: include_superseded
          in: query
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          description: Caps the result count; 0 or absent returns all matches.
          schema: { type: integer }
        - name: all_namespaces
          in: query
          description: >
            Aggregate across every namespace, ignoring the namespace header. The
            server merges all namespaces and applies limit as a single global cap
            under the requested sort, so the admin UI's "All projects" view fetches
            one response instead of one request per namespace.
          schema: { type: boolean, default: false }
        - name: namespace
          in: query
          description: >
            With all_namespaces=true, restrict the aggregate to these namespaces
            (repeatable, exact match); ignored otherwise. Lets the browser narrow
            an "All projects" listing without changing the active namespace.
          schema: { type: array, items: { type: string } }
        - name: memory_type
          in: query
          description: >
            Repeatable and/or comma-separated metadata.memory_type filter; a memory
            matches if its type is ANY of the listed values (OR). Distinct from
            "meta", which ANDs one value per key.
          schema: { type: array, items: { type: string } }
        - name: created_after
          in: query
          description: Only memories created at or after this instant.
          schema: { type: string, format: date-time }
        - name: accessed_after
          in: query
          description: Only memories last accessed at or after this instant.
          schema: { type: string, format: date-time }
        - name: sort
          in: query
          description: Column to order by. Defaults to created_at.
          schema:
            type: string
            enum: [created_at, updated_at, last_accessed_at, access_count, importance]
            default: created_at
        - name: order
          in: query
          description: Sort direction. Defaults to desc (newest / highest first).
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListResponse" }
        "400": { $ref: "#/components/responses/Error" }
    delete:
      operationId: forgetByTag
      tags: [memory]
      summary: Delete every memory in the namespace carrying a tag
      description: >
        Bulk-deletes all memories in the request's namespace that carry the
        given tag. Combined with the import provenance tag
        (import:<source>:<date>), this undoes a single bulk import in one call.
        The tag query parameter is required to avoid an accidental
        delete-everything.
      parameters:
        - { $ref: "#/components/parameters/Namespace" }
        - name: tag
          in: query
          required: true
          description: Exact tag a memory must carry to be deleted.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteByTagResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /v1/handshake:
    post:
      operationId: handshake
      tags: [memory]
      summary: Resolve namespace, identity, and behavioral settings from client-supplied project facts
      description: >
        The client→server handshake: the client sends what it knows about the
        project (git remote/toplevel/cwd, an optional agent suffix, and any
        namespace it already has an opinion about) plus its own name/version;
        the server resolves the effective namespace (an explicit pin beats
        MEMINI_NAMESPACE beats a declared value beats derivation from the
        project facts beats the caller's key-level default beats the server
        default), the caller's identity, the fully-merged ClientSettings
        (built-in defaults, overridden by the server's global defaults,
        overridden by any per-key settings), and the read-set the resolved
        namespace draws from. This is the single source of truth for
        namespace resolution — clients no longer derive or cache it locally.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HandshakeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HandshakeResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /v1/settings/defaults:
    get:
      operationId: getSettingsDefaults
      tags: [memory]
      summary: The server's global default ClientSettings
      description: >
        The defaults every key inherits from absent a per-key override.
        Admin-gated like /v1/keys (see listApiKeys): allowed for the admin
        env key, or when auth is disabled entirely (dev/bootstrap mode); a
        named table or file key gets 403 — a named caller sees its own
        merged result via /v1/handshake or /v1/self instead.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsDefaultsResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
    put:
      operationId: putSettingsDefaults
      tags: [memory]
      summary: Replace the server's global default ClientSettings
      description: >
        Admin-gated like /v1/keys (see listApiKeys): this changes what every
        key on the server inherits absent its own override. May be rejected
        when the server locks this layer read-only via its own configuration
        — a later phase's concern, not this contract's.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientSettings" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientSettings" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Namespace:
      name: X-Memini-Namespace
      in: header
      required: false
      description: Tenant/agent namespace; falls back to the server default.
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
    Tier:
      type: string
      enum: [working, episodic, semantic, procedural]
    Level:
      type: string
      enum: [explicit, deduced]
    RememberRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
        tier:
          allOf:
            - $ref: "#/components/schemas/Tier"
          description: >-
            Omit to let the server choose: the content is classified by the
            marker heuristic (a terse, unhedged decision/preference/problem
            lands in semantic/procedural, stamped metadata.tier_classified=marker),
            falling back to working. Classification never picks working
            (the default) — it only raises to semantic/procedural.
        summary: { type: string }
        tags: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: true }
        importance: { type: number, format: double, minimum: 0, maximum: 1 }
        level:
          allOf:
            - $ref: "#/components/schemas/Level"
          description: >-
            Label the derivation provenance (explicit vs deduced) at write time.
            Omit to leave unset (default, legacy rows, auto-tagged by service).
        ttl_seconds:
          type: integer
          description: Overrides the tier default TTL; negative means never expire.
        id:
          type: string
          description: Upserts an existing memory when provided.
        valid_from:
          type: string
          format: date-time
          description: >
            Start of the interval the fact was true. Defaults to now; backdate it
            to record a historical fact so time-travel (as_of) recall surfaces it.
        valid_to:
          type: string
          format: date-time
          description: >
            End of the interval the fact was true (for recording a fact that was
            true only in the past). Omit for a fact that is still true.
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: >
            Seed corroboration for a durable fact (e.g. a trusted import). Omit
            to use the default seed; ignored for short-term tiers.
        visibility:
          type: string
          description: >
            Write-side namespace scoping. Omit or "project" (default) writes to
            the request namespace itself. "personal" writes to the caller's
            home namespace (X-Memini-Home header / MEMINI_HOME on the client)
            instead — an error if no home namespace is configured. Any other
            value must name an ancestor of the request namespace, either its
            full path or an unambiguous last path segment (e.g. "acme" for
            request namespace "acme/phoenix/api"), and the write lands there
            instead. Ignored for non-durable writes: an episodic/working memory
            always stays in the request namespace regardless of visibility.
...
    MergeHint:
      type: object
      description: >
        Optional. Returned on POST /v1/memories when the write's nearest
        same-tier candidate scored at/above MEMINI_WRITE_DEDUP_SCORE and
        MEMINI_WRITE_DEDUP_ACTION is "hint". The caller can decide whether to
        merge into the near-duplicate via memory_update.
      properties:
        similar_id:
          type: string
          description: ID of the near-duplicate memory.
        similar_content:
          type: string
          description: Preview (≤200 chars) of the near-duplicate memory's content.
        score:
          type: number
          format: double
          description: Fused similarity between the new write and the near-duplicate (0..1).
        tier:
          $ref: "#/components/schemas/Tier"
...
    ListResponse:
      type: object
      required: [memories]
      properties:
        memories:
          type: array
          items: { $ref: "#/components/schemas/Memory" }
...
    DeleteByTagResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer, description: Number of memories deleted }
...
    ReadSetOrigin:
      type: string
      description: >
        Why a namespace is in the read-set: "primary" is the request
        namespace (and its subtree, when expanded), "ancestor" is a
        path-prefix cascade leg, "home" is the caller's personal namespace,
        "link" is a stored namespace link, and "call" is an explicit per-call
        namespace.
      enum: [primary, ancestor, home, link, call]
    ReadSetEntryItem:
      type: object
      required: [namespace, origin]
      properties:
        namespace: { type: string }
        origin: { $ref: "#/components/schemas/ReadSetOrigin" }
        tiers:
          type: array
          items: { $ref: "#/components/schemas/Tier" }
          description: >
            Tier restriction applied to this namespace; omitted means the
            request's own tier filter, unrestricted beyond that.
...
    Memory:
      type: object
      required:
        [
          id,
          namespace,
          tier,
          content,
          importance,
          created_at,
          updated_at,
          last_accessed_at,
          access_count,
        ]
      properties:
        id: { type: string }
        namespace: { type: string }
        tier: { $ref: "#/components/schemas/Tier" }
        level:
          allOf:
            - $ref: "#/components/schemas/Level"
          description: >-
            Derivation provenance: explicit (user-stated / heuristic) vs
            deduced (LLM-distilled). Null/omitted when the row predates the
            tag or when unset.
        content: { type: string }
        summary: { type: string }
        metadata: { type: object, additionalProperties: true }
        tags: { type: array, items: { type: string } }
        importance: { type: number, format: double }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_accessed_at: { type: string, format: date-time }
        access_count: { type: integer }
        expires_at: { type: string, format: date-time, nullable: true }
        superseded_by: { type: string, nullable: true }
        valid_from:
          type: string
          format: date-time
          nullable: true
          description: Start of the wall-clock interval the fact was true; null means open ("always, until valid_to"). Used by time-travel (as_of) recall.
        valid_to:
          type: string
          format: date-time
          nullable: true
          description: End of the interval the fact was true; null means open ("still true"). Stamped automatically when a fact is superseded.
        confidence:
          type: number
          format: double
          nullable: true
          description: Corroboration of a durable fact in [0,1]; null when not tracked.
        merge_hint:
          $ref: "#/components/schemas/MergeHint"
          description: >
            Optional. Present only on POST /v1/memories responses when the
            write's nearest same-tier candidate scored at/above
            MEMINI_WRITE_DEDUP_SCORE and MEMINI_WRITE_DEDUP_ACTION is "hint".
        auto_superseded:
          type: boolean
          description: >
            Optional. Present only on POST /v1/memories responses when the
            write's nearest same-tier candidate scored at/above
            MEMINI_WRITE_DEDUP_SCORE with MEMINI_WRITE_DEDUP_ACTION="supersede"
            and the old memory was tombstoned in the background. The caller
            still receives the new memory.
        reinforced:
          type: boolean
          description: >
            Optional. Present only on POST /v1/memories responses when the fact
            was already known and NO new memory was created: the existing memory
            was strengthened (reinforced and corroborated) and is what this
            response returns. Two paths reach it — the exact-restatement
            fingerprint fast path, and MEMINI_WRITE_DEDUP_ACTION="coalesce" when
            the incoming phrasing is not richer than the stored one. Without this
            flag a 201 would read as "created", which is exactly what did not
            happen, and the id would appear to belong to a memory the caller
            wrote when it does not.
...
    CallerIdentity:
      type: object
      required: [authenticated]
      description: Who the request authenticated as, independent of any resolved namespace.
      properties:
        authenticated: { type: boolean }
        key_name:
          type: string
          description: >
            Name of the API key that authenticated the request; absent for
            the admin key or dev mode (no named principal — see
            requireAdminOrDev's doc for the same distinction on /v1/keys).
        home:
          type: string
          description: The key's bound home namespace, if any.
        default_namespace:
          type: string
          description: The key's bound default namespace, if any.
    ClientSettings:
      type: object
      description: >
        Behavioral/injection settings, resolved by merging built-in defaults
        with any server global defaults and any per-key override. Every
        field is optional in the schema — absent means "inherit from the
        next layer down" — but a fully resolved ClientSettings (as returned
        by /v1/handshake, /v1/self, and GET /v1/settings/defaults) always
        carries every field.
      properties:
        capture_turns:
          type: boolean
          default: true
          description: Capture each user→assistant turn as episodic memory.
        session_digest:
          type: boolean
          default: true
          description: Record a session-end/stop/pre-compact digest memory.
        inline_extract:
          type: boolean
          default: true
          description: >
            Inject the directive asking the agent to save durable facts via
            memory_remember.
        auto_save:
          type: boolean
          default: true
          description: Periodically nudge the agent to persist durable memories.
        auto_save_interval:
          type: integer
          minimum: 1
          default: 10
          description: User-message interval between auto-save nudges.
        inject_briefing_pinned:
          type: integer
          minimum: 0
          default: 5
          description: Max pinned memories in the session-start briefing.
        inject_briefing_facts:
          type: integer
          minimum: 0
          default: 5
          description: Max durable semantic facts in the session-start briefing.
        inject_briefing_procedures:
          type: integer
          minimum: 0
          default: 5
          description: Max procedural how-tos in the session-start briefing.
        inject_briefing_recent:
          type: integer
          minimum: 0
          default: 3
          description: Max recent episodic entries in the session-start briefing.
        inject_briefing_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on briefing injection tokens; 0 is uncapped.
        inject_pretool_items:
          type: integer
          minimum: 0
          default: 3
          description: Max recalled items injected per file on PreToolUse.
        inject_pretool_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on per-tool injection tokens; 0 is uncapped.
        inject_pretool_min_score:
          type: number
          minimum: 0
          default: 0
          description: Floor on the fused score (>=) for a PreToolUse injection.
        inject_pretool_tools:
          type: array
          items: { type: string }
          default: [Read, Write, Edit, Glob, Grep]
          description: Tool-name allowlist that triggers a PreToolUse injection.
        inject_labels:
          type: array
          default: []
          items:
            type: string
            enum: [tier, confidence, age, reason]
            # Codegen metadata only (no wire effect): without explicit names,
            # oapi-codegen would mint a bare `Tier` constant for the "tier"
            # value, colliding with the Tier schema's generated Go type.
            x-enum-varnames:
              [InjectLabelTier, InjectLabelConfidence, InjectLabelAge, InjectLabelReason]
          description: Which annotation labels to render alongside an injected memory.
        recall:
          type: boolean
          default: true
          description: Enable recall-driven injection at all.
        capture:
          type: boolean
          default: true
          description: Enable capture (turns/digests) at all.
        recall_limit:
          type: integer
          minimum: 0
          default: 3
          description: Max memories per recall call.
        inject_recall_max_tok:
          type: integer
          minimum: 0
          default: 0
          description: Hard ceiling on recall injection tokens; 0 is uncapped.
        inject_recall_min_score:
          type: number
          minimum: 0
          default: 0
          description: Floor on the fused score (>=) for a recall injection.
        min_capture_chars:
          type: integer
          minimum: 0
          default: 0
          description: Minimum content length worth bothering to capture a turn.
        namespace_scope:
          type: string
          enum: [repo, owner_repo]
          default: repo
          description: >
            "repo" derives the namespace from the bare repo name; "owner_repo"
            disambiguates same-named repos across owners with an
            owner-repo slug (owner + "-" + repo).
        namespace_prefix:
          type: string
          default: ""
          description: Namespace path prepended ahead of the derived/declared namespace.
    SettingsDefaultsResponse:
      description: >
        The server's global default ClientSettings (fully resolved — every
        field present) plus who manages this layer, so the admin UI can render
        an env-managed layer read-only.
      allOf:
        - $ref: "#/components/schemas/ClientSettings"
        - type: object
          required: [managed_by]
          properties:
            managed_by:
              type: string
              description: >
                "api" means the layer is stored in the server's KV store and
                editable via PUT /v1/settings/defaults; "env" means it is pinned
                by the MEMINI_CLIENT_DEFAULTS environment variable, and PUT is
                refused with 409.
              enum: [api, env]
...
    HandshakeRequest:
      type: object
      required: [project]
      description: >
        Client→server handshake input: what the client currently knows about
        the project and about itself. project.cwd_basename is the only field
        every caller can always supply (even a bare directory with no git
        repo); everything else sharpens the resolution.
      properties:
        project:
          type: object
          required: [cwd_basename]
          properties:
            remote_url:
              type: string
              description: >
                Raw git remote URL, exactly as `git remote get-url origin`
                reports it — unnormalized.
            toplevel_path:
              type: string
              description: >
                Absolute git toplevel directory — the machine-local pin key
                (path:<toplevel_path>).
            toplevel_basename:
              type: string
              description: Basename of toplevel_path, for the toplevel fallback derivation.
            cwd_basename:
              type: string
              description: >
                Basename of the working directory — the last-resort
                derivation fallback, always present.
            agent:
              type: string
              description: >
                Per-agent suffix (e.g. "reviewer"), sanitized and appended as
                a nested namespace segment.
            env_namespace:
              type: string
              description: >
                The client's MEMINI_NAMESPACE, when set. Sent so a pin can
                still beat it server-side — the client cannot make that call
                itself without knowing whether a pin exists.
            declared_namespace:
              type: string
              description: >
                For gateway/integration callers with no meaningful cwd (a
                webhook relay, a CI job) to declare their namespace directly.
                Wins over derivation but not over a pin or env_namespace.
        client:
          type: object
          description: Caller identification, for logging/diagnostics only.
          properties:
            name: { type: string }
            version: { type: string }
    HandshakeResponse:
      type: object
      required:
        [namespace, namespace_source, identity, settings, settings_sources, read_set, server]
      properties:
        namespace:
          type: string
          description: The resolved namespace this caller should use for every subsequent request.
        namespace_source:
          type: string
          description: Which rule resolved `namespace`, highest precedence first.
          enum: [pin, env, declared, remote, toplevel, cwd, key_default, server_default]
        pin:
          type: object
          description: Present only when namespace_source is "pin".
          required: [key, updated_at]
          properties:
            key:
              type: string
              description: >
                The project_map key that matched
                (remote:<canonical-remote> or path:<toplevel_path>).
            note: { type: string }
            created_by:
              type: string
              description: >
                Name of the API key that created the pin, when known —
                same semantics as ProjectMapEntry.created_by (absent for
                a pin created by the admin key or in dev mode, which
                carry no named principal).
            updated_at: { type: string, format: date-time }
        identity: { $ref: "#/components/schemas/CallerIdentity" }
        settings:
          allOf:
            - $ref: "#/components/schemas/ClientSettings"
          description: Fully resolved — every field present.
        settings_sources:
          type: object
          description: Per-field provenance for `settings`, keyed by the same field names.
          additionalProperties:
            type: string
            enum: [default, global, key]
        read_set:
          type: array
          description: >
            The read-set `namespace` resolves to (same shape as GET
            /v1/namespaces/readset).
          items: { $ref: "#/components/schemas/ReadSetEntryItem" }
        server:
          type: object
          required: [version, default_namespace]
          properties:
            version: { type: string }
            default_namespace: { type: string }
...